DS-908 - Refactor Hippo Utils Service
