+++ cms/src/main/java/com/visitscotland/brxm/validator/MapPageValidator.java
@@ -2,7 +2,7 @@ package com.visitscotland.brxm.validator;
 
 import com.visitscotland.brxm.config.VsComponentManager;
 import com.visitscotland.brxm.hippobeans.General;
-import com.visitscotland.brxm.utils.HippoUtilsService;
+import com.visitscotland.brxm.services.HippoUtilsService;
 import org.hippoecm.hst.content.beans.ObjectBeanManagerException;
 import org.hippoecm.hst.content.beans.query.exceptions.QueryException;
 import org.onehippo.cms.services.validation.api.ValidationContext;
