+++ site/components/src/main/java/com/visitscotland/brxm/components/breadcrumb/VsBreadcrumbComponent.java
@@ -5,7 +5,7 @@ import com.visitscotland.brxm.hippobeans.BaseDocument;
 import com.visitscotland.brxm.hippobeans.Page;
 import com.visitscotland.brxm.config.VsComponentManager;
 import com.visitscotland.brxm.services.DocumentUtilsService;
-import com.visitscotland.brxm.utils.HippoUtilsService;
+import com.visitscotland.brxm.services.HippoUtilsService;
 import org.hippoecm.hst.content.beans.standard.HippoBean;
 import org.hippoecm.hst.core.component.HstComponentException;
 import org.hippoecm.hst.core.component.HstRequest;
