+++ site/components/src/main/java/com/visitscotland/brxm/factory/MapFactory.java
@@ -12,7 +12,7 @@ import com.visitscotland.brxm.model.*;
 import com.visitscotland.brxm.services.MapService;
 import com.visitscotland.brxm.services.ResourceBundleService;
 import com.visitscotland.brxm.utils.CMSProperties;
-import com.visitscotland.brxm.utils.HippoUtilsService;
+import com.visitscotland.brxm.services.HippoUtilsService;
 import com.visitscotland.utils.Contract;
 import org.hippoecm.hst.core.component.HstRequest;
 import org.onehippo.taxonomy.api.Category;
