+++ site/components/src/main/java/com/visitscotland/brxm/utils/InternalParameterProcessor.java
@@ -1,6 +1,7 @@
 package com.visitscotland.brxm.utils;
 
 import com.visitscotland.brxm.model.LocalizedURL;
+import com.visitscotland.brxm.services.HippoUtilsService;
 import com.visitscotland.brxm.services.ResourceBundleService;
 import com.visitscotland.utils.Contract;
 import org.apache.cxf.jaxrs.impl.UriBuilderImpl;
