+++ site/components/src/test/java/com/visitscotland/brxm/factory/ImageFactoryTest.java
@@ -15,7 +15,7 @@ import com.visitscotland.brxm.services.CommonUtilsService;
 import com.visitscotland.brxm.services.ResourceBundleService;
 import com.visitscotland.brxm.utils.CMSProperties;
 import com.visitscotland.brxm.utils.ContentLogger;
-import com.visitscotland.brxm.utils.HippoUtilsService;
+import com.visitscotland.brxm.services.HippoUtilsService;
 import org.hippoecm.hst.content.beans.ObjectBeanManagerException;
 import org.hippoecm.hst.content.beans.query.exceptions.QueryException;
 import org.junit.jupiter.api.Assertions;
