+++ site/components/src/test/java/com/visitscotland/brxm/utils/SitePropertiesTest.java
@@ -1,5 +1,6 @@
 package com.visitscotland.brxm.utils;
 
+import com.visitscotland.brxm.services.HippoUtilsService;
 import com.visitscotland.brxm.services.ResourceBundleService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
